fix: Correct signup type and add project update
The `handleSignup` function's `userData` parameter was incorrectly typed as an intersection of `Student` and `Professor`. This change corrects it to a union type, as a user can be either a student or a professor, not both simultaneously.

Additionally, a new function `handleUpdateProject` has been added to allow modification of existing projects.